requirements.txt
# PyPI packages required by the DAG.
# Upload this file alongside the composer setup script.
apache-airflow-providers-snowflake>=5.3.0
snowflake-connector-python>=3.6.0
google-cloud-aiplatform>=1.60.0
